+++ R/manCULTA-sim.R
@@ -1,0 +1,66 @@
+#' Simulation Replication
+#'
+#' @author Ivan Jacob Agaloos Pesigan
+#'
+#' @return The output is saved as an external file in `output_folder`.
+#'
+#' @inheritParams Template
+#' @export
+#' @keywords manCULTA simulation
+Sim <- function(taskid,
+                repid,
+                output_folder,
+                overwrite,
+                integrity,
+                seed) {
+  # Do not include default arguments here.
+  # All arguments should be set in `.sim/sim-args.R`.
+  # Add taskid to output_folder
+  output_folder <- file.path(
+    output_folder,
+    paste0(
+      SimProj(),
+      "-",
+      sprintf(
+        "%05d",
+        taskid
+      )
+    )
+  )
+  if (!file.exists(output_folder)) {
+    dir.create(
+      path = output_folder,
+      showWarnings = FALSE,
+      recursive = TRUE
+    )
+    .SimChMod(output_folder)
+  }
+  if (is.null(seed)) {
+    seed <- .SimSeed(
+      taskid = taskid,
+      repid = repid
+    )
+  }
+  suffix <- .SimSuffix(
+    taskid = taskid,
+    repid = repid
+  )
+  SimGenData(
+    taskid = taskid,
+    repid = repid,
+    output_folder = output_folder,
+    seed = seed,
+    suffix = suffix,
+    overwrite = overwrite,
+    integrity = integrity
+  )
+  SimFitCULTA2Profiles(
+    taskid = taskid,
+    repid = repid,
+    output_folder = output_folder,
+    seed = seed,
+    suffix = suffix,
+    overwrite = overwrite,
+    integrity = integrity
+  )
+}