+++ R/manCULTA-template.R
@@ -1,0 +1,20 @@
+#' @param taskid Positive integer.
+#'   Task ID.
+#' @param repid Positive integer.
+#'   Replication ID.
+#' @param output_folder Character string.
+#'   Output folder.
+#' @param output_type Character string.
+#'   Output type.
+#' @param suffix Character string.
+#'   Output of `manCTMed:::.SimSuffix()`.
+#' @param seed Integer.
+#'   Random seed.
+#' @param overwrite Logical.
+#'   Overwrite existing output in `output_folder`.
+#' @param integrity Logical.
+#'   If `integrity = TRUE`,
+#'   check for the output file integrity when `overwrite = FALSE`.
+#'
+#' @name Template
+NULL